change instructor and candidate service methods, add new constants

+++ src/main/java/mk/ukim/finki/drivingschoolevidencesystem/service/impl/User/CandidateServiceImpl.java
@@ -1,5 +1,6 @@
 package mk.ukim.finki.drivingschoolevidencesystem.service.impl.User;
 
+import mk.ukim.finki.drivingschoolevidencesystem.domain.constants.Constants;
 import mk.ukim.finki.drivingschoolevidencesystem.domain.dto.CandidateDTO;
 import mk.ukim.finki.drivingschoolevidencesystem.domain.exceptions.TrafficSchoolException;
 import mk.ukim.finki.drivingschoolevidencesystem.domain.models.Candidate;
@@ -10,21 +11,20 @@ import mk.ukim.finki.drivingschoolevidencesystem.security.generator.PasswordGene
 import mk.ukim.finki.drivingschoolevidencesystem.service.CandidateService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.Set;
 
 @Service("candidateService")
-public class CandidateServiceImpl implements CandidateService {
+public class CandidateServiceImpl extends UserServiceImpl implements CandidateService {
     @Autowired
     private CandidateRepository candidateRepository;
-    @Autowired
-    private RoleRepository roleRepository;
     @Autowired
     private ModelMapper modelMapper;
     @Autowired
@@ -36,17 +36,22 @@ public class CandidateServiceImpl implements CandidateService {
     @Override
     public CandidateDTO createNew(CandidateDTO candidateDTO) {
         Candidate candidate = candidateRepository.findByEmbg(candidateDTO.getEmbg());
-        if(candidate == null) {
-            String password = passwordGenerator.generateRandomPassword();
-            setCandidateData(candidate, candidateDTO);
-            candidate.setPassword(passwordEncoder.encode(password));
-            candidate = candidateRepository.save(candidate);
-            //SEND MAIL
-            candidateDTO = modelMapper.map(candidate, CandidateDTO.class);
-            return candidateDTO;
+        Role candidateRole = getRole(Constants.Role.CANDIDATE.name());
+        if(candidate == null && candidate.getRoles().contains(candidateRole)) {
+            throw new TrafficSchoolException("Candidate with embg = " + candidateDTO.getEmbg() + "already exists." );
         }
 
-        throw new TrafficSchoolException("Candidate with embg = " + candidateDTO.getEmbg() + "already exists." );
+        String password = "";
+        if(candidate == null) {
+            password = passwordGenerator.generateRandomPassword();
+            candidate.setPassword(passwordEncoder.encode(password));
+        }
+        setCandidate(candidate, candidateDTO);
+        candidate.getRoles().add(candidateRole);
+        candidate = candidateRepository.save(candidate);
+        //SEND MAIL
+        candidateDTO = modelMapper.map(candidate, CandidateDTO.class);
+        return candidateDTO;
     }
 
 
@@ -55,7 +60,7 @@ public class CandidateServiceImpl implements CandidateService {
     public CandidateDTO edit(CandidateDTO candidateDTO) {
         long id = candidateDTO.getId();
         Candidate candidate = getCandidateWithiId(id);
-        setCandidateData(candidate, candidateDTO);
+        setCandidate(candidate, candidateDTO);
         candidate = candidateRepository.save(candidate);
         candidateDTO = modelMapper.map(candidate, CandidateDTO.class);
         return candidateDTO;
@@ -65,15 +70,25 @@ public class CandidateServiceImpl implements CandidateService {
     @Transactional
     @Override
     public void remove(long id) {
+        Candidate candidate = getCandidateWithiId(id);
+        if(candidate.getRoles().size() > 1) {
+            candidate.getRoles().remove(getRole(Constants.Role.CANDIDATE.name()));
+            candidateRepository.save(candidate);
+            return;
+        }
         candidateRepository.deleteById(id);
     }
 
+    @Transactional
     @Override
-    public Collection<CandidateDTO> getAll() {
-        return null;
+    public Page<CandidateDTO> getAll(int pageNumber) {
+        Pageable pageRequest = PageRequest.of(pageNumber, Constants.Page.SIZE, Sort.by(Constants.Page.DEFAULT_SORT_PROPERTY));
+        Page<Candidate> candidatePage = candidateRepository.findAllByRoles_name(Constants.Role.CANDIDATE.name(), pageRequest);
+        Page<CandidateDTO> result = modelMapper.map(candidatePage, Page.class);
+        return result;
     }
 
-    private void setCandidateData(Candidate candidate, CandidateDTO candidateDTO) {
+    private void setCandidate(Candidate candidate, CandidateDTO candidateDTO) {
         candidate.setFirstName(candidateDTO.getFirstName());
         candidate.setLastName(candidateDTO.getLastName());
         candidate.setParentName(candidateDTO.getParentName());
@@ -83,24 +98,8 @@ public class CandidateServiceImpl implements CandidateService {
         candidate.setAddress(candidateDTO.getAddress());
         candidate.setPhoneNumber(candidateDTO.getPhoneNumber());
         candidate.setGender(candidateDTO.getGender());
-        candidate.setRoles(getRoles(candidateDTO.getRoles()));
     }
 
-    private Set<Role> getRoles(Set<String> roles) {
-        Set<Role> candidateRoles = new HashSet<>();
-        for(String name : roles) {
-            candidateRoles.add(getRole(name));
-        }
-        return candidateRoles;
-    }
-
-    @Transactional(propagation = Propagation.MANDATORY)
-    public Role getRole(String name) {
-        Role role = roleRepository.findById(name)
-                .orElseThrow(() -> new TrafficSchoolException("Role with name: " + name + " does not exist"));
-
-        return role;
-    }
 
     @Transactional(propagation = Propagation.MANDATORY)
     public Candidate getCandidateWithiId(long id) {
